Updated scripts and overall structure to better fit a python package. Still need to update README with reactome and package information, and actually add package files.
