+++ .travis.yml
@@ -11,7 +11,8 @@ install:
 script:
   - cd scripts
   # Installs pgd, uses graph conversion script, and runs PPA
-  - bash runPPA.sh ../data/Wnt wnt_ranking.txt
+  - bash setupPGD.sh
+  - bash runPPA.sh ../data/Wnt wnt_ranking.txt ../lib/pgd/ 
   - bash runNetBoxIL2.sh
   # Compare the generated IL2 output with the reference output
   - cmp ../tests/reference/il2_ranking.txt il2_ranking.txt
