+++ bin/setupPGD.sh
