+++ referencePathways/README.md
@@ -1,3 +1,7 @@
 # Reference Pathways
 
-This directory contains sets of reference pathway graphlets to use in pathway parameter advising (currently only Reactome).
+This directory contains reference pathway graphlets to use in pathway parameter advising.
+
+reactomeGraphlets.pkl is a pickled dictionary of graphlet distributions for reactome which can be used by pathway paramter advising.
+
+It is reccomended to run updateReactome.sh in the scrips directory to get the latest version of reactome.
